feat(portofolio): add update with file

+++ app/controllers/PortofolioController.php
@@ -45,11 +45,13 @@ class PortofolioController{
       $extension = substr($_FILES['gambar']['name'],strlen($_FILES['gambar']['name'])-4,strlen($_FILES['gambar']['name']));
       $file = md5($_FILES['gambar']['name']).time().$extension;
 
-      $data = PortofolioModel::detail($_GET["id"]);
-      if (file_exists('C:\laragon\www\CRUD_Injection\views\asset\img/'.$data->gambar_porto)) {
-        unlink('C:\laragon\www\CRUD_Injection\views\asset\img/'.$data->gambar_porto);
-      } 
-
+      $data = PortofolioModel::detail($_POST["id"]);
+      if ($data->gambar_porto != null) {
+        if (file_exists('C:\laragon\www\CRUD_Injection\views\asset\img/'.$data->gambar_porto)) {
+          unlink('C:\laragon\www\CRUD_Injection\views\asset\img/'.$data->gambar_porto);
+        } 
+      }
+      
       $result = move_uploaded_file($_FILES['gambar']['tmp_name'], 'C:\laragon\www\CRUD_Injection\views\asset\img/'.$file);
       $data = PortofolioModel::update($_POST["id"],$_POST["judul"],$_POST["deskripsi"],$_POST["link"],$_POST["tanggal"],$file);
       header("Location:".$url."/views/user/portofolio.php");
